+++ packages/ui-shared/src/components/ButtonLinkBox/ButtonLinkBox.docz.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const BackgroundWrapper: React.FC = ({ children }) => (
+    <div style={{ background: '#00B956' }}>{children}</div>
+);